requirements.txt
Flask
Flask-Session
psycopg2
SQLAlchemy
requests
python-dotenv
